+++ package.json
@@ -30,8 +30,8 @@
   },
   "dependencies": {
     "@sveltejs/adapter-node": "^1.0.0-next.73",
-    "moment": "^2.29.2",
-    "troi-library": "^0.0.4",
+    "moment": "^2.29.3",
+    "troi-library": "^0.0.8",
     "yup": "^0.32.11"
   },
   "type": "module"
