+++ src/lib/Troi.svelte
@@ -11,8 +11,12 @@
 
   let projects = [];
 
-  onMount(async () => {
+  const reload = async () => {
     projects = await $troiApi.getCalculationPositions();
+  };
+
+  onMount(() => {
+    reload();
   });
 </script>
 
@@ -41,6 +45,28 @@
         />
       </label>
     </div>
+    <div class="inline-block content-center pt-8">
+      <button
+        class="rounded-sm border border-blue-500 px-2 py-1 hover:bg-blue-100"
+        on:click={() => {
+          (projects = []), reload();
+        }}
+        ><svg
+          xmlns="http://www.w3.org/2000/svg"
+          class="h-6 w-6"
+          fill="none"
+          viewBox="0 0 24 24"
+          stroke="#0063eb"
+          stroke-width="2"
+        >
+          <path
+            stroke-linecap="round"
+            stroke-linejoin="round"
+            d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15"
+          />
+        </svg></button
+      >
+    </div>
   </div>
 </section>
 
@@ -48,7 +74,7 @@
   <!-- TODO: make into single component Project -->
   <section class="bg-white">
     <div class="container mx-auto pt-4 pb-2">
-      <h2 class="text-lg font-medium">
+      <h2 class="text-lg font-semibold text-gray-900">
         {project.name}
       </h2>
       <TroiTimeEntries
@@ -66,7 +92,7 @@
   <p>
     Project not showing up? Make sure it's available in Troi and marked as a
     "favorite". <br />
-    The "I'm lazy" function is not intended for actual use, it is just a fun feature.
+    The "Suggest" function is not intended for actual use, it is just a fun feature.
   </p>
 </section>
